Move FailedCommentSubmission into notice_comment.
This will require a major version bump as it could lead to data loss.
